Move ItemListContainer component to a new ItemListContainer directory
